Simplify bindings, include change note in the bindings.
I noticed that the three bindings, user info, template and template options,
are all supplied with the same dictionary of data. There is therefore no point
in passing three dictionaries if one will do. Only one set of extra_bindings is
now passed to the notifier. Also in this change: we will attempt to get the
most recent change note from portal_archivist, if it exists, and make it
available to the templates for inclusion in the notification.